refactor(projects): drop duplicate test ApiProject view model
Tests now deserialize into SheetMusic.Api.Projects.ViewModels.ApiProject. The view model gets an explicit constructor pair (parameterless for JSON, Project-based for mapping), matching the ApiPart pattern.

Part of #196
