+++ src/SheetMusic.Api/Projects/ViewModels/ApiProject.cs
@@ -3,15 +3,31 @@ using System;
 
 namespace SheetMusic.Api.Projects.ViewModels;
 
-public class ApiProject(Project project)
+public class ApiProject
 {
-    public Guid Id { get; set; } = project.Id;
+    /// <summary>
+    /// Parameterless constructor required for JSON deserialization (e.g. in integration tests).
+    /// </summary>
+    public ApiProject()
+    {
+    }
 
-    public string Name { get; set; } = project.Name;
+    public ApiProject(Project project)
+    {
+        Id = project.Id;
+        Name = project.Name;
+        Comments = project.Comments;
+        StartDate = project.StartDate;
+        EndDate = project.EndDate;
+    }
 
-    public string? Comments { get; set; } = project.Comments;
+    public Guid Id { get; set; }
 
-    public DateTime StartDate { get; set; } = project.StartDate;
+    public string Name { get; set; } = null!;
 
-    public DateTime EndDate { get; set; } = project.EndDate;
+    public string? Comments { get; set; }
+
+    public DateTime StartDate { get; set; }
+
+    public DateTime EndDate { get; set; }
 }
